// packages/bedrock-commands/src/types/parameter-type.d.ts
/** The type of command parameter */
export declare enum ParameterType {
    /** An animation key */
    animation = 0,
    /** Block identifiers */
    block = 1,
    /** Biome identifiers */
    biome = 2,
    /** Block states */
    blockStates = 3,
    /** Booleans */
    boolean = 4,
    /** causeType */
    causeType = 5,
    /** Camera Shake mode */
    cameraShakeType = 6,
    /** Camera preset */
    cameraPresets = 7,
    /** Clone mode */
    cloneMode = 8,
    /** Command start */
    command = 9,
    /** Coordinate */
    coordinate = 10,
    /** Damage cause */
    damageCause = 11,
    /** Difficulty mode */
    difficulty = 12,
    /** Dimension */
    dimension = 13,
    /** The camera easing effect */
    easing = 14,
    /** Effect identifiers */
    effect = 15,
    /** Entity identifiers */
    entity = 16,
    /** Enchant type */
    enchant = 17,
    /**  */
    executeSubcommand = 18,
    /** Event in behaviors */
    event = 19,
    /** The fill mode */
    fillMode = 20,
    /** Functions paths */
    function = 21,
    /** Float numbers */
    float = 22,
    /** Fog identifiers */
    fog = 23,
    /** Game mode */
    gamemode = 24,
    /** The type of hand: mainhand or offhand */
    handType = 25,
    /** The type of hud visibility */
    hudVisibility = 26,
    /** The type of hud element */
    hudElement = 27,
    /** Integer numbers */
    integer = 28,
    /** Integer range numbers */
    integer_range = 29,
    /** Item identifiers */
    item = 30,
    /** Json item format */
    jsonItem = 31,
    /** Json raw text */
    jsonRawText = 32,
    /** Keyword */
    keyword = 33,
    /** A loot table */
    lootTable = 34,
    /** A loot table */
    jigsaw = 35,
    /** Locates feature mode */
    locateFeature = 36,
    /** Message */
    message = 37,
    /** Mask mode */
    maskMode = 38,
    /** Mirror mode */
    mirror = 39,
    /** Music repeat mode */
    musicRepeatMode = 40,
    /** Music */
    music = 41,
    /** Objective identifiers */
    objective = 42,
    /** Old Block mode */
    oldBlockMode = 43,
    /** Operation mode */
    operation = 44,
    /** Score comparator mode */
    scoreComparator = 45,
    /** Particle identifier */
    particle = 46,
    /** Permission */
    permission = 47,
    /** Permission state */
    permissionState = 48,
    /** Recipe item */
    recipe = 49,
    /** Replace mode */
    replaceMode = 50,
    /** Ride Rules */
    rideRules = 51,
    /** Ride Fill Mode */
    ridefillMode = 52,
    /** Rotation Mode */
    rotation = 53,
    /** Save Mode */
    saveMode = 54,
    /** scan Mode */
    scanMode = 55,
    /** Selectors */
    selector = 56,
    /** Slot type */
    slotType = 57,
    /** Slot ID */
    slotID = 58,
    /** Sound identifier */
    sound = 59,
    /** Strings */
    string = 60,
    /** Structure */
    structure = 61,
    /** Structure animation  mode */
    structureAnimationMode = 62,
    /** Tag identifiers */
    tag = 63,
    /** Teleport rules */
    teleportRules = 64,
    /** Ticking area identifiers */
    tickingarea = 65,
    /** Time mode */
    time = 66,
    /** Time in ticks */
    timeInTicks = 67,
    /** Unknown */
    unknown = 68,
    /** XP */
    xp = 69
}
//# sourceMappingURL=parameter-type.d.ts.map